rate系統を専用のファイルに移動
