+++ src/rmmz/rpg/data/main/traitContainers/trait/index.ts
@@ -1,5 +1,7 @@
 export * from "./core";
 export * from "./equip";
+export * from "./rate";
+export * from "./skill";
 export * from "./trait";
 export * from "./types";
 export { formatTraitText } from "./format";
