+++ src/rmmz/rpg/data/main/traitContainers/trait/trait.ts
@@ -1,40 +1,6 @@
 import type { TraitCode } from "./core";
-import {
-  TRAIT_PARAM,
-  TRAIT_ELEMENT_RATE,
-  TRAIT_DEBUFF_RATE,
-  TRAIT_STATE_RATE,
-  TRAIT_ATTACK_ELEMENT,
-} from "./core";
+import { TRAIT_ATTACK_ELEMENT } from "./core";
 import type { Trait } from "./types";
-
-export const traitParamRate = (
-  traits: ReadonlyArray<Trait>,
-  paramId: number,
-): number => {
-  return traitPi(traits, TRAIT_PARAM, paramId);
-};
-
-export const traitElementRate = (
-  traits: ReadonlyArray<Trait>,
-  elementId: number,
-): number => {
-  return traitPi(traits, TRAIT_ELEMENT_RATE, elementId);
-};
-
-export const traitDebuffRate = (
-  traits: ReadonlyArray<Trait>,
-  paramId: number,
-): number => {
-  return traitPi(traits, TRAIT_DEBUFF_RATE, paramId);
-};
-
-export const traitStateRate = (
-  traits: ReadonlyArray<Trait>,
-  stateId: number,
-): number => {
-  return traitPi(traits, TRAIT_STATE_RATE, stateId);
-};
 
 export const traitAttackElements = (traits: ReadonlyArray<Trait>): number[] => {
   return traitSet(traits, TRAIT_ATTACK_ELEMENT);
